Fix the two asserts. Move things around. Remove createImported. Move CXXMethodBridging to it's own header.
