+++ lib/AST/Decl.cpp
@@ -7272,21 +7272,6 @@ SubscriptDecl *SubscriptDecl::create(ASTContext &Context, DeclName Name,
   return SD;
 }
 
-VarDecl *VarDecl::createImported(ASTContext &Context, DeclName Name,
-                                 SourceLoc SubscriptLoc, SourceLoc ArrowLoc,
-                                 Type ElementTy, DeclContext *Parent,
-                                 ClangNode ClangN) {
-  assert(ClangN && ElementTy);
-  auto *DeclPtr = allocateMemoryForDecl<VarDecl>(
-      Context, sizeof(VarDecl), /*includeSpaceForClangNode=*/true);
-
-  auto *const SD = ::new (DeclPtr) VarDecl(false, Introducer::Var, SourceLoc(),
-                                           Name.getBaseIdentifier(), Parent);
-  SD->setInterfaceType(ElementTy);
-  SD->setClangNode(ClangN);
-  return SD;
-}
-
 SubscriptDecl *SubscriptDecl::createImported(ASTContext &Context, DeclName Name,
                                              SourceLoc SubscriptLoc,
                                              ParameterList *Indices,
