+++ lib/Sema/TypeCheckStorage.cpp
@@ -2414,8 +2414,7 @@ IsAccessorTransparentRequest::evaluate(Evaluator &evaluator,
           break;
         }
       }
-      if (storage->hasClangNode())
-        break;
+
       if (auto subscript = dyn_cast<SubscriptDecl>(storage)) {
         break;
       }
