+++ test/Interop/Cxx/templates/not-pre-defined-class-template-module-interface.swift
@@ -10,7 +10,6 @@
 // CHECK-NEXT: struct IntWrapper {
 // CHECK-NEXT:   init()
 // CHECK-NEXT:   init(value: Int32)
-// CHECK-NEXT:   var value: Int32 { get }
 // CHECK-NEXT:   func getValue() -> Int32
 // CHECK-NEXT:   var value: Int32
 // CHECK-NEXT: }
